Encapsulate substance of tiles
Instead of working with tiles' elevations and thicknesses directly,
encapsulate them in a "substance" property that returns the tile's
composition by value. All changes to the composition happen inside
the Tile object itself.

This will allow iteration on what a tile's composition even means
without clients needing to know about it.

Biggest change is to how erosion works. When a tile was just an
elevation value, the simulation could just sum up the positive and
negative effects and apply them in one shot. Now we do it in two steps,
first eroding each tile and associating each bit of its eroding material
with its destination, and then depositing the accumulated materials at
the destination.