Make tiles consist of layers of rocks
Instead of defining a tile's dimensions with "thickness" and "elevation"
values, give it a "bottom" (the depth below sea level where the tile
begins) and a list of layers, each with a rock type and thickness; tile
thickness and elevation become properties derived from the layer
attributes.

Currently the only rock types are 'I'gneous (all starting rocks, and
layers created by subduction of ocean crust) and 'S'edimentary (those
deposited from erosion).

The most intense code is around "merging" tiles into one. This
represents the situation where multiple tiles map to a single
location as a continent moves; what was previously a simple averaging
of elevation values is now a complicated function where potentially very
dissimilar lists of layers must be combined into one "average".

The actual way this works is explained in comments in Tile.mergelayers,
and the various properties of the merge are tested in tile_test.py, but
the general goal is to preserve extrema and "unique" layers, to best
capture the diversity of an evolving terrain.

The redefinition of the thickness and elevation as properties breaks
pickling beyond migratability, which I guess is why pickling is not the
best way of implementing file serialization. In any case, older saves
are broken by this revision.