Make sediment and metamorphism preserve attributes
Metamorphism now affects each layer independently, instead of mashing
them all together, so that each layer's attributes are preserved.

Sedimentary deposits mix the attributes of their source materials'
layers according to their contributions. This is sort of complicated,
since in a given step one tile can erode to multiple destinations and a
tile can accept deposits from multiple sources.

As an example of the former, a peak might be losing 0.1 km to neighbor A
and 0.2 km to neighbor B. If the source tile's top few layers look like

    0.05 km of 50% felsic igneous rock, on top of
    0.1 km of 75% felsic igneous rock with 5% iron content, and then
    > 0.15 km of 60% felsic metamorphic rock,

then both A and B will receive deposits that are

    63.33% (0.05*50% + 0.1*75% + 0.15*60%)/0.3 felsic with
    1.67% (0.1*5%)/0.3 iron content.

Likewise, when a given tile is receiving deposits from multiple sources,
the source layers are averaged together in the same way.