Change approach to erosion
The previous erosion logic was something like: each tile erodes to all
adjacent tiles of lower elevation, by up to the elevation difference
between them, based on precipitation and glaciation.

With this change, each tile erodes to all adjacent tiles regardless of
relative elevation, based on absolute elevation, precipitation, and
temperature.

The "degree" of erosion between each source and destination is no
longer an inherently meaningful measurement of the quantity of material
transfered, but an abstract value in [0,1] representing the amount of
erosive activity between the tiles. Layers in the source tile erode
from the top down according to the toughness of each affected rock.

Erosion is also scaled by timestep, meaning that subduction-based
orogeny should stay in balance with erosion across timescales, though
this is untested.

The end effect is that continents do not appear to inevitably become
dominated by mountains as they had previously, and the new consideration
of rocks' toughness should allow the generation of Alp-like ranges
composed of the "bones" of old mountains.