Break tectonics out into its own mode
New world mode will begin with tectonics and then switch to more granular
simulations (NYI). Atmosphere and life status is now part of the save
format, and there is some expedient but not particularly well thought out
separation of building and interpreting save formats from the actual reading
and writing of the files.