Scale agricultural probability with deglaciation
As glaciers diminish, people are more likely to discover agriculture. This
is sort of questionable, since the higher population as glaciers retreat
already boosts the overall likelihood that any people will develop
agriculture. But in our history, agriculture popped up independently around
the same time, so there's some reason to think that there's something about
the end of a glacial period that makes it more likely.