Bring latlong/vector conversions back into sync
Fixes the conversion of vectors into latitude and longitude so Tile's
calculated vector matches the original input vector as used as keys in
tile maps. This makes tile selection work correctly (though still breaks
after resizing the window), and also allows using the vector property
as input into other stuff (which it currently never is).

To keep the orientation correct in the display, SphereView now does
an axis rotation for rendering, and the winding order of each face's
vertices is inverted from what it had previously been.

Still pretty dumb to have to change this everywhere instead of doing it
in a constructor or something, but that refactor put off for now.