Route rivers in terrain
Adapt the coarse river routes to the terrain scale by trying to find the next
lowest adjacent location at each step.

Sort of weird at the final step, as it can kind of circle around the bay to find
the lowest tile: would be better to know the true destination tile.