Stub nation rendering
Pick a bunch of random points as nation centers, then assign nations based on
closest such point. Some wack stand-in stuff to clean up here: choosing from
source nations randomly instead of deterministically as populations expand into
frontier, and randomized colors on map affect the RNG state (should save/restore,
or just be non-random).

The nation definition itself is also a stand-in for better logic: nations should
be smaller based on environment and civizilization (smaller in jungles and for
pre-agricultural peoples), and should be less likely to cross rivers or other
features.