Determine trading partners
Neighboring nations with different enough "resources" (currently just number of
different species within their bounds) are considered to trade together. Their
shared borders are shaded at each step, and selecting a nation highlights it and
its trade partners.

Highlighting nations reveals that a number of them are non-contiguous, which in
hindsight makes sense based on the national boundaries code. Should probably fix
this at some point.

Trade is currently limited to direct relationships between neighbors, there's no
possibility of a nation trading with neighbor A to get resources that they in turn
trade with neighbor B. I want to do something like that but with some kind of
tracking of how B's trade depends on A's, so that trade relationships don't flap
from tick to tick.