Borrow words for new import goods
Converts a number of functions to generators to find words that won't collide
with any existing ones.

Some of these throw exceptions or fall through the end to return None; in
practice it works so far just due to the number of possible words in the search
space, but I'm sure the right combination of words would cause exceptions.