Implement conquest
Nations get aggressive with neighbors that offer a combination of more natural
resources (native species) and a smaller population. They conquer a neighbor when
the population they can bring to bear exceeds the other's: they might have more
absolute population, but be involved in multiple conflicts, in which case they
are at a disadvantage.

Conquered nations currently just get completely switched over to the victorious
nation, there's no attempt to keep track of ethnicities or the conquered nation's
language or anything.