Decouple (detrouple) language/nation/state
Instead of tracking single-nation states that each speak their own language,
that becomes a starting point, and these are allowed to evolve
independently.

The new data is modeled such that each tile has a list of communities. A lot
of the properties that were previously associated with tiles or states are
now moved to the community level: each community has a racial makeup,
nationality, language, and culture, so as these communities grow in
population and move around, each tile can start to have a mix of people.

Still to do is to have national and linguistic similarities and differences
be taken into consideration when breaking up states or establishing trade,
to provide some kind of assimilation for isolated minority populations, and
to allow communities to glom together (maybe the same as assimilation?).